for #7920 moved track specific code to track_parser.py
